refactor Zoo
